+++ src/lib.rs
@@ -96,3 +96,5 @@ pub mod types;
 /// This module contains the `Wynd` struct and related server functionality
 /// for creating and managing WebSocket servers.
 pub mod wynd;
+
+pub mod handle;
